refactor(InputView, OutputView): view 패키지로 이동
